Move StackGraph to Plan package
SceptrePlan only dependency in config package is Graph. Graph is not
used anywhere else. This commit removes the dependency between plan and
config packages.
